# src/features/api/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TIDAL Hi-Fi API",
    "version": "7.0.0",
    "description": "",
    "license": {
      "name": "MIT",
      "url": "https://github.com/Mastermindzh/tidal-hifi/blob/master/LICENSE"
    },
    "contact": {
      "name": "Rick <mastermindzh> van Lieshout",
      "url": "https://www.rickvanlieshout.com"
    }
  },
...
  "paths": {
...
    "/settings/skipped-artists": {
      "get": {
        "summary": "get a list of artists that TIDAL Hi-Fi will skip if skipping is enabled",
        "tags": [
          "settings"
        ],
        "responses": {
          "200": {
            "description": "The list book.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StringArray"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Add new artists to the list of skipped artists",
        "tags": [
          "settings"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StringArray"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Ok"
          }
        }
      }
    },
    "/settings/skipped-artists/delete": {
      "post": {
        "summary": "Remove artists from the list of skipped artists",
        "tags": [
          "settings"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StringArray"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Ok"
          }
        }
      }
    },
...
    "/settings/skipped-tracks": {
      "get": {
        "summary": "get a list of track-title keywords that TIDAL Hi-Fi will skip if track-skipping is enabled (case-insensitive substring match)",
        "tags": [
          "settings"
        ],
        "responses": {
          "200": {
            "description": "The list of skipped track keywords.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StringArray"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Add new keywords to the list of skipped tracks",
        "tags": [
          "settings"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StringArray"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Ok"
          }
        }
      }
    },
    "/settings/skipped-tracks/delete": {
      "post": {
        "summary": "Remove keywords from the list of skipped tracks",
        "tags": [
          "settings"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StringArray"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Ok"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "StringArray": {
        "type": "array",
        "items": {
          "type": "string"
        },
        "example": [
          "Artist1",
          "Artist2"
        ]
      }
    }
  },
...
}